Removes default Firebase files
Removes the default `firebase.json` and `index.html` files.
Adds a basic 404 page.

This prepares the project for custom content.

Adds .gitignore file

Adds a comprehensive .gitignore file to the project.

This prevents unnecessary files and directories, such as logs, caches, and dependencies, from being included in the repository.

This helps to keep the repository clean and efficient.