Adds README for interactive presentation
Adds a README file to the project, providing information
about the interactive presentation for the Master's Thesis.

The README includes links to the live presentation
and the TFM repository, as well as instructions for
viewing the presentation locally.

Revert "Add GitHub Actions workflow for GitHub Pages deployment"

This reverts commit b4f55d21912085a96183e5b430e6166e3c0078f9.

Corrects a typo in the acknowledgements.

Fixes a typo in the acknowledgements section of the page, improving the overall polish and professionalism.

Removes unnecessary instructions

Removes the introductory instructions from the presentation guide, as they are no longer necessary and clutter the document.